Move Intake schema versions to hand-written package, use them instead of autogenerated enum
